Bit of proof cleaning with Jérémy
We can use [match goal] to avoid inlining partial sequents in the proof
scripts. Left to do is applying this technique to the proof of
[trace_of_eta], including inlining the use of [trace_of_eta'] to avoid
its monstruous statement in the first place. Note the systematic but
tedious use of [destruct] is in addition a potential target for
small-scale proof automation.